added font to cgfx

+++ src/include/exec/types.h
@@ -13,12 +13,12 @@ typedef void *				APTR;
 
 typedef signed char			INT8;
 typedef signed short		INT16;
-typedef signed int			INT32;
+typedef signed long			INT32;
 typedef signed long long int	INT64;
 
 typedef unsigned char		UINT8;
 typedef unsigned short		UINT16;
-typedef unsigned int		UINT32;
+typedef unsigned long		UINT32;
 typedef unsigned long long int	UINT64;
 
 typedef float				FLOAT;
